[tint] Fix modernize-avoid-c-style-cast, pt 3
- Non-trivial replacements for T(x)
- Uses static_cast where in Dawn I would use checked_cast because Tint
  hasn't adopted that style yet. Eventually will get back around to this
  issue with bug 538767409.

Bug: 548071770
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/334540
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/tint/cmd/common/helper.h
@@ -329,7 +329,8 @@ bool ReadStdinImpl(std::vector<T>* buffer) {
         // std::min for extra safety
         size_t items_read = std::min(chunk.size(), bytes_read / sizeof(T));
         buffer->reserve(buffer->size() + items_read);
-        std::copy(chunk.begin(), chunk.begin() + int32_t(items_read), std::back_inserter(*buffer));
+        std::copy(chunk.begin(), chunk.begin() + static_cast<ptrdiff_t>(items_read),
+                  std::back_inserter(*buffer));
     }
     return true;
 }
